separacio de CRUD finalitzada

+++ Model/esborrar.php
@@ -28,7 +28,6 @@ function esborrarArticle($id,$usuari){
             }
         }
 
-        
         // si hi ha errors, els mostrem
         if (!empty($errors)) {
             foreach ($errors as $error) {
@@ -37,34 +36,29 @@ function esborrarArticle($id,$usuari){
             return; // finalitzem l'execució si hi ha errors
         }
 
-        // mostrem el formulari de confirmacio d'eliminacio
-        if (!isset($_POST['confirmar'])) {
-            ?>
-            <h2>Estàs segur que vols eliminar l'article amb ID <?php echo htmlspecialchars($id); ?>?</h2>
-            <form method="post">
-                <input type="hidden" name="id" value="<?php echo htmlspecialchars($id); ?>">
-                <input type="hidden" name="Enviar" value="Eliminar">
-                <input type="submit" name="confirmar" value="Si">
-                <input type="submit" name="confirmar" value="No">
-            </form>
-            <?php
+       // mostrem el formulari de confirmació només si no s'ha enviat encara 
+       if (!isset($_POST['confirmar'])) {
+            include_once "../Vistes/confirmEsborrar.php";
         }
 
         // si el usuari ha confirmat la eliminació s'elimina
-        if ($_POST['confirmar'] === 'Si') {
-            $consultaEliminar = $connexio->prepare("DELETE FROM article WHERE id = :id");
-            $consultaEliminar->bindParam(':id', $id);
+        if (isset($_POST['confirmar'])) {
+            if ($_POST['confirmar'] === 'Si') {
+                // elimineu l'article de la base de dades
+                $consultaEliminar = $connexio->prepare("DELETE FROM article WHERE id = :id");
+                $consultaEliminar->bindParam(':id', $id);
 
-            if ($consultaEliminar->execute()) {
-                echo "article amb ID " . htmlspecialchars($id) . " eliminat correctament ✅";
+                if ($consultaEliminar->execute()) {
+                    echo "article amb ID " . htmlspecialchars($id) . " eliminat correctament ✅";
+                } else {
+                    echo "error al eliminar article ❌";
+                }
             } else {
-                echo "error al eliminar article ❌";
+                // si no, es cancela l'eliminació
+                echo"eliminació cancel·lada ✅";
             }
-        } else {
-            // si no, es cancela l'eliminacio
-            echo "eliminació cancel·lada ✅";
         }
-    
+
         }catch (PDOException $e) {
         // en cas d'error de connexió mostrem el missatge d'error
         echo "error de connexió: " . $e->getMessage() . " ❌";
